test(integration): float-ize foundation-model construct tests to halve OOM footprint (C/D shards)
Integration C and D shards were OOM-killed ("runner received shutdown signal")
constructing foundation-scale models at double on the 16GB runner — not the
timeouts originally assumed. These construct/predict + name/metadata tests have
dtype-invariant assertions, so switching them to float halves the construction
footprint (the proven PR #1596 lever):

- SegmentationModelSizeVariationTests (C): SAM/SAMHQ ViTHuge (~636M), SEEM,
  Mask2Former, etc. — ~5GB -> ~2.5GB at float, fits the runner.
- DiffusionExtendedIntegrationTests (D): ~50 foundation diffusion model
  construction tests (Sora, WanVideo, Flux, SDXL, PixArt, ...). The 2
  DiffusionModelOptions<double> config tests keep double (they assert exact
  values like 0.001).
- LayoutAwareDocumentTests (D): LayoutLM(v2/v3)/Donut/DocOwl construct + predict
  + metadata-name tests; the 0.5 image-fill literal becomes 0.5f.

Assertions are ParameterCount>0 / output shape / metadata names / config values
— all dtype-invariant, so no test is weakened. The largest models that still
exceed the runner even at float are addressed separately via lazy construction.

+++ tests/AiDotNet.Tests/IntegrationTests/Diffusion/DiffusionExtendedIntegrationTests.cs
@@ -73,21 +73,21 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task StableDiffusion2Model_Construction()
     {
-        var model = new StableDiffusion2Model<double>();
+        var model = new StableDiffusion2Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task StableDiffusion3Model_Construction()
     {
-        var model = new StableDiffusion3Model<double>();
+        var model = new StableDiffusion3Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task SDXLModel_Construction()
     {
-        var model = new SDXLModel<double>();
+        var model = new SDXLModel<float>();
         Assert.NotNull(model);
     }
 
@@ -115,7 +115,7 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task DallE3Model_Construction()
     {
-        var model = new DallE3Model<double>();
+        var model = new DallE3Model<float>();
         Assert.NotNull(model);
     }
 
@@ -142,7 +142,7 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task Imagen2Model_Construction()
     {
-        var model = new Imagen2Model<double>();
+        var model = new Imagen2Model<float>();
         Assert.NotNull(model);
     }
 
@@ -154,13 +154,13 @@ public class DiffusionExtendedIntegrationTests
         // diffusion tests to FP32`): Kandinsky 3.0's paper-faithful defaults
         // (prior + decoder UNets + MoVQ-GAN VAE) instantiate ~16 GB of weight
         // tensors at FP64, which OOMs the CI runner — confirmed locally:
-        // `new KandinskyModel<double>()` throws OutOfMemoryException at
+        // `new KandinskyModel<float>()` throws OutOfMemoryException at
         // ConvolutionalLayer.EnsureInitialized while constructing the decoder
         // UNet's level-3 ResBlocks, while `new KandinskyModel<float>()` succeeds
         // in ~45s on the same machine. FP32 is the production-canonical
         // precision for SD/Kandinsky paper checkpoints (FP32 master / FP16
         // working). The other Construction smoke tests in this file currently
-        // fit at FP64 in CI's specific scheduling — they're left at <double>
+        // fit at FP64 in CI's specific scheduling — they're left at <float>
         // until CI demonstrates they too need migration, so the surface area
         // of this fix matches the actual CI failure exactly.
         var model = new KandinskyModel<float>();
@@ -170,70 +170,70 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task PixArtModel_Construction()
     {
-        var model = new PixArtModel<double>();
+        var model = new PixArtModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task PixArtSigmaModel_Construction()
     {
-        var model = new PixArtSigmaModel<double>();
+        var model = new PixArtSigmaModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task PixArtDeltaModel_Construction()
     {
-        var model = new PixArtDeltaModel<double>();
+        var model = new PixArtDeltaModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Flux1Model_Construction()
     {
-        var model = new Flux1Model<double>();
+        var model = new Flux1Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task HunyuanDiTModel_Construction()
     {
-        var model = new HunyuanDiTModel<double>();
+        var model = new HunyuanDiTModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task KolorsModel_Construction()
     {
-        var model = new KolorsModel<double>();
+        var model = new KolorsModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task PlaygroundV25Model_Construction()
     {
-        var model = new PlaygroundV25Model<double>();
+        var model = new PlaygroundV25Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task RAPHAELModel_Construction()
     {
-        var model = new RAPHAELModel<double>();
+        var model = new RAPHAELModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task EDiffIModel_Construction()
     {
-        var model = new EDiffIModel<double>();
+        var model = new EDiffIModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task OmniGenModel_Construction()
     {
-        var model = new OmniGenModel<double>();
+        var model = new OmniGenModel<float>();
         Assert.NotNull(model);
     }
 
@@ -244,119 +244,119 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task AnimateDiffModel_Construction()
     {
-        var model = new AnimateDiffModel<double>();
+        var model = new AnimateDiffModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task CogVideoModel_Construction()
     {
-        var model = new CogVideoModel<double>();
+        var model = new CogVideoModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task StableVideoDiffusionModel_Construction()
     {
-        var model = new StableVideoDiffusion<double>();
+        var model = new StableVideoDiffusion<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task OpenSoraModel_Construction()
     {
-        var model = new OpenSoraModel<double>();
+        var model = new OpenSoraModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task SoraModel_Construction()
     {
-        var model = new SoraModel<double>();
+        var model = new SoraModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task MakeAVideoModel_Construction()
     {
-        var model = new MakeAVideoModel<double>();
+        var model = new MakeAVideoModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task RunwayGenModel_Construction()
     {
-        var model = new RunwayGenModel<double>();
+        var model = new RunwayGenModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task VideoCrafterModel_Construction()
     {
-        var model = new VideoCrafterModel<double>();
+        var model = new VideoCrafterModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task LatteModel_Construction()
     {
-        var model = new LatteModel<double>();
+        var model = new LatteModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task ModelScopeT2VModel_Construction()
     {
-        var model = new ModelScopeT2VModel<double>();
+        var model = new ModelScopeT2VModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task LTXVideoModel_Construction()
     {
-        var model = new LTXVideoModel<double>();
+        var model = new LTXVideoModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task LuminaT2XModel_Construction()
     {
-        var model = new AiDotNet.Diffusion.TextToImage.LuminaT2XModel<double>();
+        var model = new AiDotNet.Diffusion.TextToImage.LuminaT2XModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Mochi1Model_Construction()
     {
-        var model = new Mochi1Model<double>();
+        var model = new Mochi1Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task HunyuanVideoModel_Construction()
     {
-        var model = new HunyuanVideoModel<double>();
+        var model = new HunyuanVideoModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task VeoModel_Construction()
     {
-        var model = new VeoModel<double>();
+        var model = new VeoModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task KlingModel_Construction()
     {
-        var model = new KlingModel<double>();
+        var model = new KlingModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task WanVideoModel_Construction()
     {
-        var model = new WanVideoModel<double>();
+        var model = new WanVideoModel<float>();
         Assert.NotNull(model);
     }
 
@@ -367,21 +367,21 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task AudioLDMModel_Construction()
     {
-        var model = new AudioLDMModel<double>();
+        var model = new AudioLDMModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task AudioLDM2Model_Construction()
     {
-        var model = new AudioLDM2Model<double>();
+        var model = new AudioLDM2Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task DiffWaveModel_Construction()
     {
-        var model = new DiffWaveModel<double>();
+        var model = new DiffWaveModel<float>();
         Assert.NotNull(model);
     }
 
@@ -401,49 +401,49 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task RiffusionModel_Construction()
     {
-        var model = new RiffusionModel<double>();
+        var model = new RiffusionModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task StableAudioModel_Construction()
     {
-        var model = new StableAudioModel<double>();
+        var model = new StableAudioModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task BarkModel_Construction()
     {
-        var model = new BarkModel<double>();
+        var model = new BarkModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task SoundStormModel_Construction()
     {
-        var model = new SoundStormModel<double>();
+        var model = new SoundStormModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task JEN1Model_Construction()
     {
-        var model = new JEN1Model<double>();
+        var model = new JEN1Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task UdioModel_Construction()
     {
-        var model = new UdioModel<double>();
+        var model = new UdioModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task VoiceCraftModel_Construction()
     {
-        var model = new VoiceCraftModel<double>();
+        var model = new VoiceCraftModel<float>();
         Assert.NotNull(model);
     }
 
@@ -454,84 +454,84 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task ShapEModel_Construction()
     {
-        var model = new ShapEModel<double>();
+        var model = new ShapEModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task PointEModel_Construction()
     {
-        var model = new PointEModel<double>();
+        var model = new PointEModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Zero123Model_Construction()
     {
-        var model = new Zero123Model<double>();
+        var model = new Zero123Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Magic3DModel_Construction()
     {
-        var model = new Magic3DModel<double>();
+        var model = new Magic3DModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task TripoSRModel_Construction()
     {
-        var model = new TripoSRModel<double>();
+        var model = new TripoSRModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task One2345Model_Construction()
     {
-        var model = new One2345Model<double>();
+        var model = new One2345Model<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task MVDreamModel_Construction()
     {
-        var model = new MVDreamModel<double>();
+        var model = new MVDreamModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task SyncDreamerModel_Construction()
     {
-        var model = new SyncDreamerModel<double>();
+        var model = new SyncDreamerModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Wonder3DModel_Construction()
     {
-        var model = new Wonder3DModel<double>();
+        var model = new Wonder3DModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task MeshyModel_Construction()
     {
-        var model = new MeshyModel<double>();
+        var model = new MeshyModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task LGMModel_Construction()
     {
-        var model = new LGMModel<double>();
+        var model = new LGMModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task Instant3DModel_Construction()
     {
-        var model = new Instant3DModel<double>();
+        var model = new Instant3DModel<float>();
         Assert.NotNull(model);
     }
 
@@ -542,63 +542,63 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task ControlNetModel_Construction()
     {
-        var model = new ControlNetModel<double>();
+        var model = new ControlNetModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task ControlNetXSModel_Construction()
     {
-        var model = new ControlNetXSModel<double>();
+        var model = new ControlNetXSModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task ControlNetUnionModel_Construction()
     {
-        var model = new ControlNetUnionModel<double>();
+        var model = new ControlNetUnionModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task T2IAdapterModel_Construction()
     {
-        var model = new T2IAdapterModel<double>();
+        var model = new T2IAdapterModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task IPAdapterModel_Construction()
     {
-        var model = new IPAdapterModel<double>();
+        var model = new IPAdapterModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task IPAdapterFaceIDModel_Construction()
     {
-        var model = new IPAdapterFaceIDModel<double>();
+        var model = new IPAdapterFaceIDModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task InstantIDModel_Construction()
     {
-        var model = new InstantIDModel<double>();
+        var model = new InstantIDModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task PhotoMakerModel_Construction()
     {
-        var model = new PhotoMakerModel<double>();
+        var model = new PhotoMakerModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task UniControlNetModel_Construction()
     {
-        var model = new UniControlNetModel<double>();
+        var model = new UniControlNetModel<float>();
         Assert.NotNull(model);
     }
 
@@ -609,28 +609,28 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task ConsistencyModel_Construction()
     {
-        var model = new ConsistencyModel<double>();
+        var model = new ConsistencyModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task LatentConsistencyModel_Construction()
     {
-        var model = new LatentConsistencyModel<double>();
+        var model = new LatentConsistencyModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task SDTurboModel_Construction()
     {
-        var model = new SDTurboModel<double>();
+        var model = new SDTurboModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task AuraFlowModel_Construction()
     {
-        var model = new AuraFlowModel<double>();
+        var model = new AuraFlowModel<float>();
         Assert.NotNull(model);
     }
 
@@ -641,7 +641,7 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task SDUpscalerModel_Construction()
     {
-        var model = new SDUpscalerModel<double>();
+        var model = new SDUpscalerModel<float>();
         Assert.NotNull(model);
     }
 
@@ -652,14 +652,14 @@ public class DiffusionExtendedIntegrationTests
     [Fact(Timeout = 120000)]
     public async Task BlendedDiffusionModel_Construction()
     {
-        var model = new BlendedDiffusionModel<double>();
+        var model = new BlendedDiffusionModel<float>();
         Assert.NotNull(model);
     }
 
     [Fact(Timeout = 120000)]
     public async Task DiffEditModel_Construction()
     {
-        var model = new DiffEditModel<double>();
+        var model = new DiffEditModel<float>();
         Assert.NotNull(model);
     }
